Update DoctorList.java
The table needs the number of doctors to display. To add doctors to the table, "isCellEditable" may need to be moved or changed. Other than that, the DoctorList GUI doesn't need to be edited at all unless it looks terrible on Windows or something.

I made sure to make it idiot-proof, so it can probably be recycled for other GUIs.