Create RetrieveInfoWindow.java
If I put too much in this GUI, you can swap one block of code with another on the bottom. JTables have a lot of difficulty wrapping text in cells, so it's not feasible to put reports/medical histories in cells, unless they're extremely short.

The second column is editable, since the HSP staff are supposed to be admins now.